dnsdist: Better handling of spurious wake-ups, interrupted calls
If we were woken up by the multiplexer but have actually nothing to read,
or the call to `recv()` is interrupted, we do not want to stop listening
for the health-check response event. It is also not useful to log about
it, even at "verbose health-check" level.
Note that we would have logged previously, so this kind of event would
not have gone unnoticed anyway.