Fix building our fuzzing targets from a dist tarball
Until now all of our fuzzing targets where built with the authoritative
server, even though one of them is specific to dnsdist.
This made it easy to build all of them at once, especially for OSS-Fuzz
and CI-Fuzz, but had the unfortunate drawback of pulling several
dnsdist-specific files into the main pdns/ directory for no good
reason. It also prevented building the fuzzing targets from a dist
tarball/directory.
This commit moves the dnsdist-specific fuzzing target to the dnsdist
build process, and ensure that the standalone_fuzz_target_runner.cc
file is part of the dist tarball, making it possible to build the
fuzzing targets from the dist.
It does not move the dnsdist-specific files to the pdns/dnsdistdist/
directory yet because this would conflict with existing PRs.