Meson: Add transient dependencies for the lmdb backend
On e.g. MacOS or anywhere where include headers are not in the usual
place (e.g. /usr/include), the include directories are not added when building the backend
files. This means a transient include wouldn't be found. In this case this include is
lmdbbackend.cc -> lmdb-safe.hh -> lmdb.h