dnsdist: Fix unit tests for incoming DoH w/ nghttp2
These tests were failing on EL with LTO enabled, and passing
everywhere else. It turns out that we did not properly reset
the `s_processQuery` hack that we used in these tests to
simulate the policy decision (rules and actions), and thus
inherited what the last test set it to instead of the default
(dropping queries), which was very unexpected.