dnsdist: Fix a race in TLS certificate generation in the tests
The OCSP and TLS tests are generating new TLS certificate and key,
so there was a time window for a certificate/key mismatch if a test
is starting right at the moment the OCSP and TLS tests are generating
new material. This commit fixes it by using different file names for
these tests.