dnsdist: Set proper levels when logging messages
This PR re-evaluates the log level of all messages in the DNSdist
code base. The rules are simple:
- 'Error' should only be used when a critical service will not function
  properly
- 'Warning' should be used when a non-essential feature may not be
  working as expected
- 'Info' is for everything that should not go unnoticed but also cannot
  be triggered by an outside attacker, otherwise we might be flooding
  our logs
- 'Verbose' can be used for everything else.

There are a few exceptions, something we log at the 'info' level for
events that might be happening a lot, but only when told to do so,
like for example with verbose health-check reporting.