Fix a spuriously failing recursorcache_cc test: reset globals (indirectly) used
For SyncRes tests we have a general mechanism setting the globals
before the test are run, but the non-syncres tests do not have that, while
they still use some globals.  In this particular case, the test would
fail if the last SyncRes test run before was setting
SyncRes::s_locked_ttlperc.

While there, pass the time to the prune functions. This avoids
potential timing issues for some tests.