Add endbr64 instructions in the right spots for OpenBSD/amd64
OpenBSD has IBT (aka control flow enforcement) on amd64, for processor
models that support it. But it turns out that the existing endbr64
instructions in the boost assembly for context switching are not
enough: there needs to be endbr64 instructions right after the
jump_fcontext() call, as indirect branches end up there.