Add and document a `localwho()` function for LUA records, so you can change your answer based on which address or port a query came in from.
Can be used like this:

```
berthub.eu	300	IN	LUA	A ";if(localwho:getPort() == 5300) then return '217.100.190.174'; else return '86.82.68.237'; end"
```

This, combined with some forwarding/iptables, means that queries that came in on 217.100.190.174 also get 217.100.190.174 as an answer. And same for 86.82.68.237.
In this way you can do some simple DNS-based dual homing.