rec: introduce command to set aggressive NSEC cache size
To avoid race conditions creating and destroying the cache, we do
not allow setting the cache size if it is not enabled from startup.
Also, setting the cache size to 0 does not disable the cache, it
just makes it mostly ineffective (entries can still appear in it
and get used until the purge task is run).

Fixes #13265.