Meson: Define BOOST_CONTAINER_USE_STD_EXCEPTIONS on the command-line
This replaces the definition coming through config.h because boost container
implementations seem to trip up on certain build configurations (e.g. debugoptimized).

Additionally, it seems that when this is defined through config.h, boost does not export
the symbols related to boost exceptions to the user but is perfectly happy to throw such
exceptions.