dnsdist: Fall back to libcrypto for authenticated encryption
We used to fall back to plain-text for console communications when
libsodium was not available, which was not great. Now that we are
also using the authenticated encryption module to secure our QUIC
tokens, let's fall back to OpenSSL's Chacha20 Poly 1305
implementation instead.
Note that, unfortunately, both implementations are not compatible
so the console communication format will be different depending on
whether libsodium is available. I believe this is still better than
plain-text :)