dnsdist: Clean up the Lua objects before exiting
When code coverage is enabled, we try to call `exit()` instead of our
usual use of `_exit()` to get more accurate coverage. It does however
trigger a race condition between the destruction of Lua objects and
the Lua context(s) they belong to, so we try to explicitly clean up
the objects before calling `exit()`.