dnsdist: make sure we do not allocate 16-byte aligned objects through lua(jit)
luajit aligns only to 8 bytes by default, and some objects require
16 byte alignment.

Fixes #13766

Note that the static assert in LuaContext.hpp is commented out in
one case.  This trips on some platforms, but does not seem to be
harmful right now.

The fundamental solution remains the have luajit agree with C++ on
minimal alignment of its allocators.