add support for domains in also-notify
I am running a powerdns setup in Kubernetes with dynamical provisioned
secondaries which I want to notify. Kubernetes provides through it's
services domains, which can be used to discover all the available
secondaries.

This commit adds support to lookup and resolve all the IPs of a domain
in also-notify, using the already existing `ns.lookup()` function.