rec: allow recursor.conf file to contain YAML
This should us to work around the packaging issues discussed in #13935.
THe idea is that modify the parsing so that .conf files also *may* contain YAML.

The search for a config file then becomes:

1. Try read recuror.yml if it exists. If valid, done. If it is invalid punt.
2. Try read recursor.conf as YAML. If it is valid, done.
3. If it is invalid, try to read as old-style.

This means that the status of recursor.conf as a config file does not change.

This allows us to install a default YAML config into recursor.conf for new installs.
Of course we leave recursor.conf (and recursor.yml) alone for existing installs.

This is a draft. I will add docs and packaging changes after this is deemed
the way to proceed.