dnsdist: Use atomic variables for the per-protocol latencies
While we do not care about a race in the sense that it's OK if we
miss/overwrite a value from time to time, we still need to prevent
an interleaved write from two threads which could technically result
in a garbage value. Using atomics also makes it easier for TSAN to
understand what's going on.