dnsdist: Disable `SONAME` support in Quiche
Since 0.22.0 Quiche sets a proper `SONAME`

See https://github.com/cloudflare/quiche/pull/1769

but it does not matter in our case since we install
the Quiche library in such a way (libdnsdist-quiche.so)
that we are the only user, and it will always be updated
with DNSdist. Keeping it makes our life significantly harder
since several packaging tools look a the `SONAME`.