rec: fix a race in the dnstap regression tests
A test might finish "too early". In that case not all workers are
fully setup yet, and leak sanitizer crashes due to a stack out of
bounds.

One case fixed by introducing a sleep (for a test to see if the
dnstap queue is empty), and one actually querying an existing
out-of-band domain, which was enough for me to get all threads
properly setup. Previously it would check a domain and get
connnection refused, as this test does not setup auths at all.
The assumption is that that causes a "too fast exit" of rec.