rec: release ref to key in MTasker::waitEvent
If we do not do that, only the next call of sendEvent will release
the shared pointer to the TCPConnectionHanlder in PacketID. This
might delay cleaning TCP objects up on not-so-busy threads.

Fixes #13422, thought it can still take a while on idle recursors,
as the housekeeping function is not called often in those cases.