dnsdist: add support for labels on custom metrics
This adds support for prometheus labels on custom metrics. Changes the
metrics maps to hold a map of labels to metric values for each metric.
Metrics without labels have an empty string label combination value.

Adds an optional options table as the last parameter of `incMetric`,
`decMetric`, `setMetric` and `getMetric`, which can be used to set
labels by using the `label` key or to set the previously used parameter
(either `step` or `value`).

Closes: #13359