pdnsutil {add-record,delete-rrset}: Don't append ZONE if NAME ends with . or ZONE
If a NAME ends with a . it is to be understood as an absolute name and
appending the zone is not intuitive then.

Note this changes behaviour for calls like:

	pdnsutil --config-dir=configs/auth add-record example.net . NS 1.2.3.4

which added the NS record to the zone's apex before and is likely an
error now.

Also make both

	pdnsutil --config-dir=configs/auth add-record example.net www.example.net A 1.2.3.5
	pdnsutil --config-dir=configs/auth add-record example.net www A 1.2.3.5

add www.example.net. to the example.net zone.

Closes: https://github.com/PowerDNS/pdns/issues/8595