rec: check bounds of rcode stats counter index (safe right now)
Safe right now as LWResult::d_rcode gets assigned from the 4 bit
rcode in the header.  But that might change one day. I'd rather
make LWResult::d_rcode an uint8_t, but that causes a conflict with
the OOB resolving code that does not make a difference between res
and d_rcode.