dnsdist: Use 65535 instead of 255 to block all types via eBPF
Our eBPF filtering code used to treat the `255` (`ANY`) qtype as a
special value meaning to block queries for all types. Unfortunately
it prevented blocking queries for the `ANY` type only, which is a
legitimate use-case.
After this commit `255` is no longer a special value, and `65535`
(a reserved value) is used instead to represent all types.

This is a breaking change that should NOT be backported.