dnsdist: Properly export FFI functions when building with `meson`
Some FFI functions are not used by the `dnsdist` binary, but they are
exported nevertheless to be usable by Lua code. The visibility of
such functions is correctly set to `public`, but our `meson` build
was putting these functions in an intermediate, static library then
linking that library in the `dnsdist` binary, and unfortunately
symbol visibility is ignored by the linker in this case.
So this commit moves the FFI function out of the shared library and
link them directly into the `dnsdist` binary, making them available
to Lua again.