dnsdist: Upgrade clang to 19 in our CI
To get rid of the `WARNING: Symbolizer buffer too small` warning,
which is caused by big backtraces, and fixed in clang >= 15 by
using a dynamic symbolizer buffer size.
I only upgrade it for dnsdist because of a compatibility issue between
`libfaketime` and the ASAN implementation in recent versions of `clang`:

https://github.com/wolfcw/libfaketime/issues/365

It seems to be fixed in the `libfaketime` repository. There has not
been any release since the fix, but I guess we could compile from
a more recent commit.