Simplify !a.isPartOf(b) && a != b constructs.
By design, a.isPartOf(a) is always true.
Therefore, if a and b compare equal, !a.isPartOf(b) and a != b are both
false and the result of the expression is false, but also
!a.isPartOf(b).

If not, then a != b is true and the result of the expression is
!a.isPartOf(b).

Boolean algebra is hard, let's go shopping.